+++ CHANGELOG.md
@@ -4,6 +4,7 @@ based on [Semantic Versioning](http://semver.org/)
 
 *2.0.0* 2014-04-02
 
++ add increment attributes in concept
 + add util/text_transformers.dart and use it for codes and labels
 + add Telephone, Name, Description and Money attribute types
 + derive default labels, if null, from code and codes